backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic[email]==2.7.1
pydantic-settings==2.2.1
motor==3.4.0
pymongo==4.6.3
beanie==1.25.0
rapidfuzz==3.9.3
jellyfish==1.0.4
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
regex==2024.4.16
Unidecode==1.3.8
Faker==24.11.0
PyYAML==6.0.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
loguru==0.7.2
httpx==0.27.0
python-multipart==0.0.9
pytest==8.2.0
pytest-asyncio==0.23.6
joblib==1.4.2
bcrypt==3.2.2
